fix sdk dependencies, ts is dev and make pg a peer
Typescript is a development dependency, not a runtime dependency.

pg as a peer feels like a better fit for the sdk, usually apps will
have pg already installed since they are using it for the rest of their app.
It also has the added "flair" that the library shows up as 0 dependencies on npm.